separate op funcs from analysis funcs
